requirements.txt

streamlit
pandas
numpy
requests
matplotlib
seaborn
python-dotenv